nadakki-google-ads-mvp/requirements.txt
# NADAKKI AI Suite - Google Ads Integration
# Core Dependencies

# Google Ads
google-ads>=21.0.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0

# Database
asyncpg>=0.28.0
sqlalchemy>=2.0.0

# Security
cryptography>=41.0.0
python-jose>=3.3.0

# HTTP Client
httpx>=0.24.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Observability
prometheus-client>=0.17.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dateutil>=2.8.0
